Moved code from original uilding blocks to new pipeline.
git-svn-id: http://usg.lofar.org/svn/code/trunk@156 e3a606d5-b720-0410-b6ad-cb58f335abf8

+++ src/CR-Pipeline/implement/Detection/PeakSearch.cc
@@ -21,6 +21,7 @@
 #include <fstream>
 #include <sstream>
 #include <iostream>
+#include <cmath>
 #include <Detection/PeakSearch.h>
 #include <Detection/SourceFit.h>
 
@@ -97,7 +98,7 @@ namespace CR { // Namespace CR -- begin
   }
   
   PeakList PeakSearch::findPeaks(Data* d) {
-    if (d->length() < (uint)(1 << blocksize_) + abs(delay_)) {
+    if (d->length() < (uint)(1 << blocksize_) + abs((double)delay_)) {
       cerr << "  ERROR: Dataset contains too little samples to run findPeaks algorithm." << endl;
       return PeakList();
     }
@@ -120,7 +121,7 @@ namespace CR { // Namespace CR -- begin
 	  sdv += (abs(d->data(j))-avg)*(abs(d->data(j))-avg);
 	}
 	sdv >>= blocksize_;
-	sdv = (uint)sqrt((double)sdv);
+	sdv = (uint)std::sqrt((double)sdv);
 	k = 0;
 	limit = avg + threshold_*sdv;
       }
